+++ domain/src/main/java/com/example/domain/model/LikeRequest.kt
@@ -2,6 +2,6 @@ package com.example.domain.model
 
 data class LikeRequest(
     val createdAt: String,
-    val memberId: Int,
+    val memberId: Long,
     val studioId: Int
 )
