+++ domain/src/main/java/com/example/domain/repository/reservation/ReservationRepository.kt
@@ -6,10 +6,10 @@ import com.example.domain.model.Reservation
 import com.example.domain.model.ReservationDetail
 
 interface ReservationRepository {
-    suspend fun getReservationsByMemberId(memberId: Int): List<Reservation>
+    suspend fun getReservationsByMemberId(memberId: Long): List<Reservation>
     suspend fun getReservationDetailByReservationId(reservationId: Int): ReservationDetail
-    suspend fun getCompletedReservationByMemberId(memberId: Int): List<Reservation>
-    suspend fun deleteReservationByReservationId(reservationId: Int, memberId: Int)
+    suspend fun getCompletedReservationByMemberId(memberId: Long): List<Reservation>
+    suspend fun deleteReservationByReservationId(reservationId: Int, memberId: Long)
     suspend fun getReservationTime(studioId: Int, date: String): AvailableReservationTime
     suspend fun makeReservation(receipt: NewReservation)
 }
